Added viewport initial-scale=1 to the home and tag template files
Reason is to use the 1024 device px of the iPad when it's in landscape
mode, not blown up 768 css px. This also means that the +800px "wide"
layout is shown instead of the mobile layout. Only problem is that
pesky orientation change bug. https://gist.github.com/1568180 might be
a potential fix.