Fix width issues and other hacks.
For whatever reason, this wasn't scaling itself automagically on my devices.
Setting the svg width and height to window.innerWidth/Height did the trick.

Also, at least on my galaxy tab--the buttons won't work unless you flip from
portrait to landscape orientation. That's weird. As a sort of hack, I've moved
the player over to the right--this seems to kick it. The other working solution
was to alert('blah'). But that just seems wrong. I don't know if this is a race
condition or what.