Compiler changed to support new panel strategy.
The panel.js, panel.css, modernizr.js will all be added in the markup.
options.js gets generated on the fly based on config.yml (in the
layouts/options.js template).

I also removed the HTML minifying because it wasn't doing anything
interesting. However, JSDOM seems to mangle the DOM of the demos so
that's not entirely solved.