Adding in a way to patch odin and emberwind.
This happens only to the deploy folders, post copy. I'm not sure
exactly the best way to test this--I tried creating a simple module
in a fixtures directory and applying a patch then asserting the difference.
But that didn't work, probably because it's 11:43 on a Friday evening.

Note that if we ever update the submodules, we need to make sure
the patches we have in the shinydemos-misc repo will work.